fix: renaming the feature to a range.
